requirements.txt
numpy
pandas
scikit-learn
xgboost
optuna
matplotlib
rdkit
shap
pyyaml
joblib
